[+/*] Added Menu and Party Screen.
Implemented player menu and party screen with basic functionality:

-The menu appears and disappears.
-You can step through all the options in the menu.
-You can select the pokemon option to transition into the pokemon party scene.
-Inside the party screen you can step through all the pokemon.

NOTE: A singleton was implemented to make player and scene manager refs easier to get but there is a crash once you change scene with those refs.

Moreover the tall grass animation does not seem to work when changing the scene.