[*] Party updates with player Pokemon.
Now the party screen gets updated with the Pokemon assigned to the player in the ready function.

- It changes the level text, party and gender sprites.

- Implemented a randomize Pokemon information function in Pokemon.cs

NOTE: we will need to retrieve some information from a JSON or DB to display some valueable information such as the Pokemon name, or access to the base stats of the Pokemon to calculate more info.